fixed duplicate code for postings, created new postingUtils file for shared posting utility functions. Implemented the Ended postings to be displayed at the bottom of the list regardless of sorting field.
