+++ server/src/api/routes/utils/postingList.ts
@@ -1,3 +1,5 @@
+import { sql } from 'kysely';
+
 import { getSingleQueryValue } from './queryValue.ts';
 
 export type PostingSortDir = 'asc' | 'desc';
@@ -208,7 +210,16 @@ export const sortPostingsBySharedSort = <T extends PostingSortLike>(
   sortBy: SharedPostingSortBy,
   sortDir: PostingSortDir,
 ): T[] => {
+  const todayKey = normalizeDateKey(new Date())!;
+
   return [...postings].sort((left, right) => {
+    // Always push ended postings to the bottom regardless of sort field
+    const leftEndKey = normalizeDateKey(left.end_date);
+    const rightEndKey = normalizeDateKey(right.end_date);
+    const leftEnded = leftEndKey !== null && leftEndKey < todayKey ? 1 : 0;
+    const rightEnded = rightEndKey !== null && rightEndKey < todayKey ? 1 : 0;
+    if (leftEnded !== rightEnded) return leftEnded - rightEnded;
+
     switch (sortBy) {
       case 'created_at':
       {
@@ -260,18 +271,24 @@ export const applySharedPostingSort = <Q extends PostingQueryLike>(
   sortBy: SharedPostingSortBy,
   sortDir: PostingSortDir,
 ): Q => {
+  // Always push ended postings to the bottom regardless of sort field
+  const withEndedLast = query.orderBy(
+    sql`CASE WHEN posting.end_date < CURRENT_DATE THEN 1 ELSE 0 END`,
+    'asc',
+  ) as Q;
+
   switch (sortBy) {
     case 'created_at':
-      return query
+      return withEndedLast
         .orderBy('posting.created_at', sortDir)
         .orderBy('posting.id', sortDir) as Q;
     case 'title':
-      return query
+      return withEndedLast
         .orderBy('posting.title', sortDir)
         .orderBy('posting.id', sortDir) as Q;
     case 'start_date':
     default:
-      return query
+      return withEndedLast
         .orderBy('posting.start_date', sortDir)
         .orderBy('posting.start_time', sortDir) as Q;
   }
